requirements.txt
customtkinter
Pillow
ultralytics
python-dotenv
customtkinter
Pillow